Support codegen generic on server side
