+++ include/grpcpp/impl/codegen/service_type.h
@@ -139,6 +139,14 @@ class Service {
     methods_[index].reset();
   }
 
+  void MarkMethodCodegenGeneric(int index) {
+    GPR_CODEGEN_ASSERT(methods_[index]->handler() != nullptr &&
+                       "Cannot mark the method as 'codegen generic' because it "
+                       "has already been "
+                       "marked as 'async' or 'generic'.");
+    methods_[index]->ResetHandler();
+  }
+
   void MarkMethodStreamed(int index, internal::MethodHandler* streamed_method) {
     GPR_CODEGEN_ASSERT(methods_[index] && methods_[index]->handler() &&
                        "Cannot mark an async or generic method Streamed");
